requirements.txt
boto3
Flask
mysql-connector-python
requests